script_db/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Django==3.0
djangorestframework==3.12.4
flake8==4.0.1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
isort==5.10.1
kombu==5.2.4
mccabe==0.6.1
oauth2client==4.1.3
prompt-toolkit==3.0.29
protobuf==4.21.1
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
rabbitmq==0.2.0
requests==2.28.0
rsa==4.8
six==1.16.0
sqlparse==0.3.1
toml==0.10.2
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.5
